Allow pin import from CSV file
- format:
  padname,pinname,direction,gatename

The CSV line can have 0 or more fields. A CSV line is only used if a
padname field is given.

- direction:
Nominally usees the names in Pin::direction_lut. For user convenience,
exact direction string representation is relaxed a bit to allow any
casing and to use space instead of '_'.

- gatename:
If empty then the gate falls back to the program default.