Sheet References (#387)
* Allow copying of sheet UUID

A context menu has been added to the sheet box. By right clicking the name of a sheet, the context menu will give users the option to copy the UUID of the selected sheet.

* Dynamically replace page references in texts on schematic

A reference to another page can now be added to a by writing "$sheetref:" followed by the UUID of the corresponding page. For display purposes, this will be replaced by the number of the referenced page in square brackets.

* Enable HUD display of sheet references

Sheet references are displayed in HUD as links to the sheet with proper title. Right now, text fields can either contain links or sheet references. When both are present, clickable links are only displayed for the sheet reference.