Disable outline creation in pdf export for podofo version 0.9.5
There is a bug that existed in podofo from revision 1775 to 1826. Because of this, only version 0.9.5 should be affected. With this bug it is not possible to create a PdfOutlineItem.

Since e.g. Ubuntu 18.04 LTS ships with this version podofo, it is not possible to export PDFs from horizon on Ubuntu 18.04 with libpodofo from the official repositories. The best remedy for this is to skip creating an outline since all the other steps run fine on podofo 0.9.5.