Fix draw connection line on pins on package origin.
When a pin is directly on the origin of a package (e.g. pin 1 on
every generic pin-header), the draw connection tool wouldn't work on
said pin because it'd receive the BoardPackage instead.

It seems like the ideal fix would be for tools to declare what object
types they're expecting to receive for an action, and have the tool
infrastructure do the disambiguation, but for now, when this happens,
scan the package for a pad on the origin to use instead.