Add shorted pads rule
These rules allow the user to specify components/parts that should be
considered as shortened for airwires. For example a 0 Ohm bridge
required for 1 layer boards should be considered as a connection and
thus shouldn't result in an airwire spanning across it.

This resulted in a app_version bump of the board.