core: catch tool_begin/tool_update reentrance
This can happen if a tool does something in update or begin that
eventually calls the update method again.

Examples include  Gtk::Clipboard::get()->request_contents which may or
may immediately invoke the received callback if the the requested data
is available in-process and imp->set_layer().

While this hasn't shown any adverse effects so far, it's still
surprising and might lead to bugs in the future.

To fix this issue, store ToolArgs from tool_update if we're in
tool_begin or tool_update. The imp will then check if there are any
pending updates after a tool has returned.

Here's how that's supposed to work:

 - tool_update/begin is called because the user did something
 - the tool then does something that results in tool_update being called
 - tool_update notices this (tool_state != ToolState::NONE) and puts
   the args into pending_tool_args and returns pretending the tool did
nothing
 - the imp's tool_process gets called to process the return value of
   tool_update which should do nothing
 - core->get_pending_tool_args() returns nothing since we're still in
   tool_update
 - eventually, the very first tool_update/begin returns and tool_process
   gets called again
 - after having processed its response, it checks if there are any
   pending tool args, calls tool_update with them and proceses their
results