Revert "imp: get monitor filenames from caching pool"
This reverts commit 490255334ff18fcaea4529b2b480d2d6cd1e1652.

changelog: Bugfixes/Editors: don't accidentally cache dependencies when opening items in a project pool

Can't remember why I did this change back then, but here's my
understanding for why it's wrong:

This only makes a difference when the current pool is a project pool.

The problem is that calling get_filename on the caching pool will copy
items from the source pool to the project pool.

In a project this should
be a no-op since the item was alredy copied to the project pool by a
call to get_filename in the tool or so.

When editing/viewing an item in a project pool however, the imp will
call this method to get the item's dependencies and the project pool
will copy them over to its cache, which is not what we want since
looking at items included into a project pool should not cause its
dependencies to be cached.