imp package: add bottom-up STEP projection
When drawing packages, while the top-down STEP projection is useful, it
can in some cases be just entirely silly...  for example QFN parts in a
top down view are just the package outline.

Add a bottom-up projection where you actually see useful things.

Note that this is not a rotation, it's a reversal of the Z axis.  This
doesn't seem to be possible in gp_Ax2 since it turns around the
handedness of the coordinate system.  Hence some transform shenanigans
on this.

Bonus points for gp_Trsf trying to be clever and converting the (1,1,-1)
matrix into a "(-1,-1,1) + scale -1" matrix, the scale part of which is
then later ignored by default in HLRAlgo_Projector...  so suddenly it's
an XY flip :'(